Update production models test
